app/requirements.txt
aiomysql==0.1.1
alembic==1.8.1
anyio==3.6.1
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.0
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.1.0
fastapi==0.70.0
fastapi-csrf-protect==0.2.1
greenlet==0.4.13
gunicorn==20.1.0
h11==0.13.0
hpy==0.0.3
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Mako==1.2.3
MarkupSafe==2.1.1
motor==2.5.1
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
protobuf==3.20.1
py==1.11.0
pydantic==1.10.1
PyJWT==2.3.0
pymongo==3.12.3
PyMySQL==1.0.2
pyparsing==3.0.9
PyPika==0.48.9
pytest==7.1.3
pytest-asyncio==0.19.0
pytest-parallel==0.1.1
python-decouple==3.3
readline==6.2.4.1
requests==2.28.1
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.16.0
tblib==1.7.0
tomli==2.0.1
typing_extensions==4.3.0
ulid-py==1.1.0
urllib3==1.26.12
uvicorn==0.14.0
